+++ src/db/models/account.js
@@ -3,26 +3,31 @@ const mongoose = require("mongoose");
 const schema = new mongoose.Schema({
     id: String,
     username: String,
-    password: String,
     email: String,
     role: String, // owner || admin || standard
     secret: String,
-    passwordless: Boolean,
-    totp: {
-        enabled: Boolean,
-        secret: String,
-        backupCodes: [String],
+    loginMethods: {
+        password: String,
+        webAuthn: Boolean,
     },
-    webauthn: {
-        lastChallenge: String,
-        authenticators: [
-            {
-                id: Buffer,
-                credentialPublicKey: Buffer,
-                counter: Number,
-                name: String,
-            },
-        ],
+    twoFactorAuthentication: {
+        enabled: Boolean,
+        totp: {
+            secret: String,
+            backupCodes: [String],
+            enabled: Boolean,
+        },
+        webAuthn: {
+            lastChallenge: String,
+            authenticators: [
+                {
+                    id: Buffer,
+                    credentialPublicKey: Buffer,
+                    counter: Number,
+                    name: String,
+                },
+            ],
+        },
     },
 });